Prevent flickering when showing captureRetake view on Android
Until now, when captureRetake view was shown on Android the Camera view and captureRetake view
were swapped in render function according to some boolean expression. This is very problematic on Android
as it creates an artifact (The previous screen would flash briefly) due to the way SurfaceView works on Android.
This commit changes things a bit by keeping the camera view in place and showing the captureRetake when needed
on top of the camera.